+++ base/deprecated.jl
@@ -1591,6 +1591,8 @@ end
 @deprecate readstring(filename::AbstractString) read(filename, String)
 @deprecate readstring(cmd::AbstractCmd) read(cmd, String)
 
+@deprecate momenttype(::Type{T}) where {T} typeof((zero(T)*zero(T) + zero(T)*zero(T))/2) false
+
 # END 0.7 deprecations
 
 # BEGIN 1.0 deprecations